+++ lib/src/widgets/menu_entry_widget.dart
@@ -4,20 +4,21 @@ import 'package:provider/provider.dart';
 
 import '../core/models/context_menu_entry.dart';
 import '../core/models/context_menu_item.dart';
+import '../core/utils/default_menu_shortcuts.dart';
 import 'context_menu_state.dart';
 
 /// A widget that represents a single item in a context menu.
 ///
 /// This widget is used internally by the `ContextMenu` contextMenu.
-class MenuEntry<T> extends StatefulWidget {
+class MenuEntryWidget<T> extends StatefulWidget {
   final ContextMenuEntry entry;
-  const MenuEntry({super.key, required this.entry});
+  const MenuEntryWidget({super.key, required this.entry});
 
   @override
-  State<MenuEntry<T>> createState() => _MenuEntryState<T>();
+  State<MenuEntryWidget<T>> createState() => _MenuEntryWidgetState<T>();
 }
 
-class _MenuEntryState<T> extends State<MenuEntry<T>> {
+class _MenuEntryWidgetState<T> extends State<MenuEntryWidget<T>> {
   late final FocusNode focusNode;
 
   ContextMenuEntry get entry => widget.entry;
@@ -25,23 +26,42 @@ class _MenuEntryState<T> extends State<MenuEntry<T>> {
   @override
   void initState() {
     focusNode = FocusNode();
+
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
     final menuState = context.read<ContextMenuState>();
+
     return MouseRegion(
-        onEnter: (event) => _onMouseEnter(context, event, menuState),
-        onExit: (event) => widget.entry.onMouseExit(event, menuState),
-        onHover: (event) => widget.entry.onMouseHover(event, menuState),
-        child: () {
+      onEnter: (event) => _onMouseEnter(context, event, menuState),
+      onExit: (event) => widget.entry.onMouseExit(event, menuState),
+      onHover: (event) => _onMouseHover(event, menuState),
+      child: Builder(
+        builder: (_) {
           if (entry is ContextMenuItem) {
-            return (entry as ContextMenuItem).builder(context, menuState, focusNode);
+            final item = entry as ContextMenuItem;
+            final shortcuts = defaultMenuShortcuts(context, item, menuState);
+            shortcuts.addAll(menuState.shortcuts);
+            return CallbackShortcuts(
+              bindings: shortcuts,
+              child: Focus(
+                focusNode: item.isFocusMaintained ? null : focusNode,
+                onFocusChange: (value) {
+                  if (value) {
+                    _ensureFocused(item, menuState, focusNode);
+                  }
+                },
+                child: item.builder(context, menuState, focusNode),
+              ),
+            );
           } else {
             return entry.builder(context, menuState);
           }
-        }());
+        },
+      ),
+    );
   }
 
   /// Handles the mouse enter event for the context menu entry.
@@ -69,5 +89,18 @@ class _MenuEntryState<T> extends State<MenuEntry<T>> {
       menuState.setFocusedEntry(item);
     }
     widget.entry.onMouseEnter(event, menuState);
+  }
+
+  _onMouseHover(PointerHoverEvent event, ContextMenuState menuState) {
+    if (menuState.isFocused(entry)) {
+      _ensureFocused(entry, menuState, focusNode);
+    }
+    widget.entry.onMouseHover(event, menuState);
+  }
+
+  void _ensureFocused(
+      ContextMenuEntry entry, ContextMenuState menuState, FocusNode focusNode) {
+    menuState.focusScopeNode.requestFocus(focusNode);
+    menuState.setFocusedEntry(entry);
   }
 }