+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 * updated example project.
 * added `ContextMenuRegion` widget to show a context menu when a user right-clicks or long-presses on a widget.
 * improvements and fixes.
+* Globally applied focus node highlighting and keyboard shortcuts for menu items.
+* Added `shortcuts` property in the `ContextMenu` class for custom context menu shortcuts.
 * updated `README.md`.